chore: satisfy backend lint baselines
