+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/worker_pool.go
@@ -2,7 +2,6 @@ package runtime
 
 import (
 	"context"
-	"crypto/sha256"
 	"errors"
 	"fmt"
 	"strings"
@@ -959,37 +958,4 @@ func appendTranscriptEntry(
 		Role:    role,
 		Summary: content,
 	})
-}
-
-func appendPromptTranscriptEntry(
-	transcript *[]workers.AgentRunTranscriptEntry,
-	role string,
-	content string,
-) {
-	content = strings.TrimSpace(content)
-	if content == "" {
-		return
-	}
-	hash := sha256.Sum256([]byte(content))
-	summary := fmt.Sprintf("sha256:%x", hash)
-	if effort := safeReasoningEffortClassification(content); effort != "" {
-		// Keep the bounded, non-prompt configuration fact that operators use to
-		// explain runner behavior without re-exposing the authored prompt.
-		summary += " (Reasoning effort: " + effort + ")"
-	}
-	appendTranscriptEntry(transcript, role, summary)
-}
-
-func safeReasoningEffortClassification(content string) string {
-	for _, line := range strings.Split(content, "\n") {
-		key, value, found := strings.Cut(line, ":")
-		if !found || !strings.EqualFold(strings.TrimSpace(key), "reasoning effort") {
-			continue
-		}
-		switch value = strings.ToLower(strings.TrimSpace(value)); value {
-		case "none", "minimal", "low", "medium", "high", "max":
-			return value
-		}
-	}
-	return ""
 }
