+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/worker_session_control_targets.go
@@ -2,6 +2,7 @@ package runtime
 
 import (
 	"context"
+	"crypto/sha256"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -969,4 +970,85 @@ func recordDispatchWorkerSessionAssociation(
 		return
 	}
 	ledger.RecordDispatchWorkerSessionAssociation(tick, dispatchID, workerSessionID, requestID, eventTime)
+}
+
+func recordedObservationFromFact(fact recordedDispatchObservation, clock factory.Clock) workersessions.Observation {
+	state := fact.state
+	if state == "" {
+		state = workersessions.StateStarting
+	}
+	observation := workersessions.Observation{
+		WorkerSessionID:          fact.workerSessionID,
+		Model:                    cloneRecordedString(fact.model),
+		ReasoningEffort:          cloneRecordedString(fact.reasoningEffort),
+		ProviderSessionAvailable: fact.provider != nil && fact.provider.ID != "",
+		WorkIDs:                  append([]string(nil), fact.workIDs...),
+		TurnID:                   fact.turnID,
+		AttemptID:                fact.dispatchID,
+		State:                    state,
+		DurationBasis:            workersessions.DurationBasisUnavailable,
+		Transcript:               workersessions.TranscriptAvailabilityUnavailable,
+	}
+	if fact.provider != nil {
+		observation.ProviderSession = providerSessionRef(*fact.provider)
+	}
+	if !fact.startedAt.IsZero() {
+		started := fact.startedAt.UTC()
+		observation.StartedAt = &started
+		if fact.endedAt != nil {
+			ended := fact.endedAt.UTC()
+			observation.EndedAt = &ended
+			duration := ended.Sub(started)
+			if duration < 0 {
+				duration = 0
+			}
+			observation.Duration = &duration
+			observation.DurationBasis = workersessions.DurationBasisRecordedTimestamps
+		} else if !state.Terminal() && clock != nil {
+			duration := clock.Now().Sub(started)
+			if duration < 0 {
+				duration = 0
+			}
+			observation.Duration = &duration
+			observation.DurationBasis = workersessions.DurationBasisActiveClock
+		}
+	}
+	if fact.failure != nil {
+		failure := *fact.failure
+		observation.Failure = &failure
+	}
+	return observation
+}
+
+func appendPromptTranscriptEntry(
+	transcript *[]workers.AgentRunTranscriptEntry,
+	role string,
+	content string,
+) {
+	content = strings.TrimSpace(content)
+	if content == "" {
+		return
+	}
+	hash := sha256.Sum256([]byte(content))
+	summary := fmt.Sprintf("sha256:%x", hash)
+	if effort := safeReasoningEffortClassification(content); effort != "" {
+		// Keep the bounded, non-prompt configuration fact that operators use to
+		// explain runner behavior without re-exposing the authored prompt.
+		summary += " (Reasoning effort: " + effort + ")"
+	}
+	appendTranscriptEntry(transcript, role, summary)
+}
+
+func safeReasoningEffortClassification(content string) string {
+	for _, line := range strings.Split(content, "\n") {
+		key, value, found := strings.Cut(line, ":")
+		if !found || !strings.EqualFold(strings.TrimSpace(key), "reasoning effort") {
+			continue
+		}
+		switch value = strings.ToLower(strings.TrimSpace(value)); value {
+		case "none", "minimal", "low", "medium", "high", "max":
+			return value
+		}
+	}
+	return ""
 }
