+++ pkg/services/workers/internal/service/execute_test_helpers_test.go
@@ -2,8 +2,10 @@ package service_test
 
 import (
 	"context"
+	"fmt"
 	"strconv"
 	"sync"
+	"sync/atomic"
 	"testing"
 	"time"
 
@@ -70,6 +72,108 @@ func mustExecuteServiceWithContentMaterializer(
 		t.Fatalf("NewWithProviderOverrideAndContentMaterializer() error = %v", err)
 	}
 	return service
+}
+
+func TestExecuteMaterializesWorkContentBeforeRunnerAndCleansItUp(t *testing.T) {
+	t.Parallel()
+
+	const materializedPath = "C:/attempt-content/image.png"
+	var materializedURL string
+	var cleanupCalls atomic.Int32
+	var runnerCalls atomic.Int32
+	service := mustExecuteServiceWithContentMaterializer(
+		t,
+		&stubRunner{execute: func(
+			_ context.Context,
+			request workers.RunnerExecutionRequest,
+		) (workers.RunnerExecutionResult, error) {
+			runnerCalls.Add(1)
+			assertMaterializedContentRequest(t, request, materializedPath)
+			return workers.RunnerExecutionResult{Content: "content accepted"}, nil
+		}},
+		work.ContentMaterializeFunc(func(
+			_ context.Context,
+			rawURL string,
+		) (string, work.ContentCleanup, error) {
+			materializedURL = rawURL
+			return materializedPath, func() { cleanupCalls.Add(1) }, nil
+		}),
+	)
+
+	request := validExecuteRequest("dispatch-content", "attempt-content")
+	request.Target.Environment.WorkingDirectory = "C:/workspace"
+	request.Input.Work = []workers.WorkInput{{
+		Name: "submitted-image",
+		Content: []work.WorkContentPart{{
+			Type:        work.WorkContentPartTypeImage,
+			URL:         "file://submitted/image.png",
+			ContentType: "image/png",
+			Metadata:    map[string]any{"origin": "submitted-work"},
+		}},
+	}}
+
+	result, err := service.Execute(context.Background(), request)
+	if err != nil {
+		t.Fatalf("Execute() error = %v", err)
+	}
+	if result.Outcome != workers.ExecutionOutcomeAccepted {
+		t.Fatalf("outcome = %q, want ACCEPTED", result.Outcome)
+	}
+	wantMaterializedURL, err := work.ResolveDispatchContentURL(
+		"C:/workspace",
+		"file://submitted/image.png",
+	)
+	if err != nil {
+		t.Fatalf("resolve expected materialized URL: %v", err)
+	}
+	if materializedURL != wantMaterializedURL {
+		t.Fatalf("materialized URL = %q, want %q", materializedURL, wantMaterializedURL)
+	}
+	if runnerCalls.Load() != 1 || cleanupCalls.Load() != 1 {
+		t.Fatalf("runner calls = %d, cleanup calls = %d, want one each", runnerCalls.Load(), cleanupCalls.Load())
+	}
+	original := request.Input.Work[0].Content[0]
+	if original.URL != "file://submitted/image.png" || original.File != "" {
+		t.Fatalf("caller content mutated = %#v, want original URL-only content", original)
+	}
+}
+
+func assertMaterializedContentRequest(
+	t *testing.T,
+	request workers.RunnerExecutionRequest,
+	wantPath string,
+) {
+	t.Helper()
+	if len(request.InputTokens) != 1 {
+		t.Fatalf("input token count = %d, want 1", len(request.InputTokens))
+	}
+	token, ok := request.InputTokens[0].(workers.Token)
+	if !ok || len(token.Color.Content) != 1 {
+		t.Fatalf("runner input token = %#v, want one typed content token", request.InputTokens[0])
+	}
+	part := token.Color.Content[0]
+	if part.File != wantPath || part.URL != "" {
+		t.Fatalf("runner content = %#v, want materialized file without source URL", part)
+	}
+	if part.ContentType != "image/png" || part.Metadata["origin"] != "submitted-work" {
+		t.Fatalf("runner content metadata = %#v, want content type and metadata preserved", part.Metadata)
+	}
+}
+
+type unsafeRemoteContentMaterializer struct{}
+
+func (*unsafeRemoteContentMaterializer) MaterializeContentURL(
+	context.Context,
+	string,
+) (string, work.ContentCleanup, error) {
+	return "", nil, fmt.Errorf("materialization should not be reached")
+}
+
+func (*unsafeRemoteContentMaterializer) ValidateContentURLSafety(
+	context.Context,
+	string,
+) error {
+	return fmt.Errorf("reject private target: %w", work.ErrUnsafeContentURL)
 }
 
 type recordingWorktree struct {
