+++ pkg/wire/recordings_providers.go
@@ -3,6 +3,7 @@ package wire
 import (
 	"fmt"
 
+	processcontract "github.com/portpowered/infinite-you/pkg/initializer/process"
 	platformclock "github.com/portpowered/infinite-you/pkg/platform/clock"
 	"github.com/portpowered/infinite-you/pkg/platform/logging"
 	platformreplay "github.com/portpowered/infinite-you/pkg/platform/replay"
@@ -10,6 +11,7 @@ import (
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factorydefinitionswire "github.com/portpowered/infinite-you/pkg/services/factory_definitions/wire"
 	factorysessionwire "github.com/portpowered/infinite-you/pkg/services/factory_sessions/wire"
+	operatorsettings "github.com/portpowered/infinite-you/pkg/services/operator_settings"
 	recordings "github.com/portpowered/infinite-you/pkg/services/recordings"
 	recordingscli "github.com/portpowered/infinite-you/pkg/services/recordings/transports/cli"
 	recordingswire "github.com/portpowered/infinite-you/pkg/services/recordings/wire"
@@ -80,4 +82,75 @@ func provideFactorySessionReplayInputs(
 	logger logging.Logger,
 ) recordings.ReplayInputLoader {
 	return recordingswire.NewReplayInputLoader(recordings.RecordingReadFile(replayFiles), loadReplay, logger)
+}
+
+// provideRecordingsProjectionCapability carries the already-composed
+// Recordings root across the neutral initializer boundary. The root package
+// reifies only the projection methods needed by callers.
+func provideRecordingsProjectionCapability(
+	service recordings.Service,
+) processcontract.RecordingsProjectionCapability {
+	var projection recordings.ProjectionService
+	if opening, ok := service.(recordings.RuntimeOpening); ok {
+		projection = opening.Projection()
+	}
+	return recordingsProjectionCapability{service: service, projection: projection}
+}
+
+type recordingsProjectionCapability struct {
+	service    recordings.Service
+	projection recordings.ProjectionService
+}
+
+func (capability recordingsProjectionCapability) RecordingsProjection() any {
+	return recordingsProjection{
+		service:    capability.service,
+		projection: capability.projection,
+	}
+}
+
+type recordingsProjection struct {
+	service    recordings.Service
+	projection recordings.ProjectionService
+}
+
+func (projection recordingsProjection) ReconstructWorldState(
+	request recordings.ReconstructWorldStateRequest,
+) (recordings.ReconstructWorldStateResult, error) {
+	return projection.service.ReconstructWorldState(request)
+}
+
+func (projection recordingsProjection) QueryWorkstationRequests(
+	request recordings.WorkstationRequestsQueryRequest,
+) (recordings.WorkstationRequestsQueryResult, error) {
+	return projection.service.QueryWorkstationRequests(request)
+}
+
+func (projection recordingsProjection) ReconstructFactoryWorldState(
+	events []recordings.FactoryEvent,
+	selectedTick int,
+) (recordings.FactoryWorldState, error) {
+	return projection.projection.ReconstructFactoryWorldState(events, selectedTick)
+}
+
+func (projection recordingsProjection) ProjectWorkstationRequests(
+	world recordings.FactoryWorldState,
+) recordings.WorkstationFactoryWorldWorkstationRequestProjectionSlice {
+	return projection.projection.ProjectWorkstationRequests(world)
+}
+
+// provideOperatorSettingsCapability carries the already-composed Operator
+// Settings root across the neutral initializer boundary for public bindings.
+func provideOperatorSettingsCapability(
+	service operatorsettings.Service,
+) processcontract.OperatorSettingsCapability {
+	return operatorSettingsCapability{service: service}
+}
+
+type operatorSettingsCapability struct {
+	service operatorsettings.Service
+}
+
+func (capability operatorSettingsCapability) OperatorSettings() any {
+	return capability.service
 }
